+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New constraints `WangLi`, `SUAL`, and `AssignmentDependent` constraints.
 - Activation and Deactivation for new constraints in `main_cli.py` as CLI arguments.
 - README file for the `assets/instances` directory.
+- ignore `examples{20-29}.txt` files in the `.gitignore`.
 
 ### Changed
 - Updated the GUI `views.py` and `controllers.py` to fit in to our new solver solution (`ortools-fast.py` but now in a dynamic and versatile way).
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated CLI `main_cli.py` script to properly handle solver, solution, and its verification.
 - Added statistics visibility for UNSAT results in the statistics tab of GUI.
 - Introduced Separation of Concerns (SOC) for our `VariableManager` class.
+- Better large complex instance generated content for instance examples.
 
 ### Refactoring
 - Introduced `InstanceParser` class for better readability for use for global function `parse_instance_file`.